% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wglmmModularised-package.R
\docType{package}
\name{wglmmModularised-package}
\alias{wglmmModularised}
\alias{wglmmModularised-package}
\title{wglmmModularised: Modularised Weighted Generalized Linear Mixed Models}
\description{
Estimates modularised generalized linear mixed models and linear mixed models with transformed dependent variable under a flexible random effects structure and -- if desired -- with survey weights.
}
\author{
\strong{Maintainer}: Akash Kumar Yadav \email{s4akyada@uni-trier.de}

}
\keyword{internal}
